… specs/E01/F01/T05/context.md
… specs/E01/F01/T05/spec.md
… specs/E01/F01/deliverables/docs/PERFORMANCE_OPTIMIZATION.md